level2/train_container/requirements.txt
google-cloud-storage
scikit-learn